requirements.txt
fastapi==0.115.2
loguru==0.7.2
matplotlib==3.9.2
numpy==2.1.2
pandas==2.2.3
passlib==1.7.4
Pillow==11.0.0
pydantic==2.9.2
pydantic_settings==2.5.2
pyecharts==2.0.6
python_jose==3.3.0
Requests==2.32.3
SQLAlchemy==2.0.34
SQLAlchemy==2.0.25
streamlit==1.36.0
streamlit_chat==0.1.1
streamlit_cookies_manager==0.2.0
streamlit_echarts==0.4.0
streamlit_option_menu==0.4.0
uvicorn==0.32.0
yfinance==0.2.44
